Use ruff's default line-length
We have already disabled the line-length warning from `ruff check`,
so this directive already have no effect.

We're about to replace `black` with `ruff format`, and in order for the
latter to behave as similarly to the former as possible, we want to
configure both with the same line-length. Both tools default to line-
length = 88, so that is what we want to use as well. Simply remove the
directive to achieve this.